Db reimplement / PublishRequest.API implemented / Login-Logout methods

+++ Backend/Services.API/Controllers/PublishRequestController.cs
@@ -1,11 +1,13 @@
 ﻿using App.Library;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using Services.API.Data;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace Services.API.Controllers
@@ -18,7 +20,7 @@ namespace Services.API.Controllers
         private readonly ILogger<PublishRequestController> _logger;
         private readonly AuthContext _context;
 
-        public PublishRequestController( ILogger<PublishRequestController> logger, AuthContext context)
+        public PublishRequestController( ILogger<PublishRequestController> logger, AuthContext context, IHttpContextAccessor httpContextAccessor)
         {
             _logger = logger;
             _context = context;
@@ -29,7 +31,14 @@ namespace Services.API.Controllers
         [Route("[action]")]
         public List<PublishRequestDetail> GetAll()
         {
-            var request_list = _context.PublishRequestDetails.ToList();
+            var request_list = _context.PublishRequestDetails.Select(x => x)
+                .Include(x => x.Game).ToList();
+
+            foreach (var req in request_list)
+            {
+                var genres = _context.GenreDetails.Where(x => x.Games.Contains(req.Game)).ToList();
+                req.Game.Genres = genres;
+            }
 
             List<PublishRequestDetail> toReturn = request_list;
             return toReturn;
@@ -39,7 +48,14 @@ namespace Services.API.Controllers
         [Route("[action]")]
         public GenericResponse<PublishRequestDetail> GetById(int id)
         {
-            var request = _context.PublishRequestDetails.Find(id);
+            var request = _context.PublishRequestDetails.Include(x => x.Game)
+                .ThenInclude(x => x.Genres).Where(r => r.ID == id).FirstOrDefault();
+
+            //var game = _context.GameDetails.Include(x => x.Genres);
+
+            //var game = _context.GameDetails.Where(x => x.ID == request.GameId).FirstOrDefault();
+            //var g = _context.GameDetails.Where(x => x.ID == request.GameId).SelectMany(c => c.Genres).ToList();
+            //request.Game = game;
             GenericResponse<PublishRequestDetail> toReturn = new GenericResponse<PublishRequestDetail> { Response = request, Code = ResponseCode.OK }; 
             return toReturn;
         }
@@ -48,9 +64,31 @@ namespace Services.API.Controllers
         [Route("[action]")]
         public GenericResponse<PublishRequestDetail> Post(GameAndGenres data)
         {
-            //data.Game.Genres = data.Genres;
+            var game = new GameDetail
+            {
+                AvailableAgeScala = data.Game.AvailableAgeScala,
+                ChildrenSuitable = data.Game.ChildrenSuitable,
+                Description = data.Game.Description,
+                GameApk = data.Game.GameApk,
+                GameName = data.Game.GameName,
+                GamePrice = data.Game.GamePrice,
+                ImageUrl = data.Game.ImageUrl,
+                LanguageOption = data.Game.LanguageOption,
+                Publisher = data.Game.Publisher,
+                Rating = 0
+            };
+            PublishRequestDetail gameRequest = new PublishRequestDetail
+            {
+                UserId = data.UserId,
+                RequestDate = DateTime.Now
+            };
 
-            PublishRequestDetail gameRequest = new PublishRequestDetail { Game = data.Game, RequestDate = DateTime.Now, UserId = data.UserId };
+            gameRequest.Game = game;
+
+            var genres = _context.GenreDetails.Where(x => data.GenreIds.Contains(x.GenreID)).ToList();
+
+            gameRequest.Game.Genres = genres;
+
             _context.PublishRequestDetails.Add(gameRequest);
             _context.SaveChanges();
 
@@ -60,15 +98,26 @@ namespace Services.API.Controllers
         }
 
 
+
+        //Sadece admin yapacak
         [HttpDelete]
         [Route("[action]")]
         public GenericResponse<PublishRequestDetail> Delete(int id)
         {
-            var toDelete = _context.PublishRequestDetails.Find(id);
-            _context.PublishRequestDetails.Remove(toDelete);
-            _context.SaveChanges();
 
-            var toReturn = new GenericResponse<PublishRequestDetail> { Response = toDelete, Code = ResponseCode.OK };
+            var role = _context.httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.Role);
+            Console.WriteLine(role);
+            var toDelete = _context.PublishRequestDetails.Find(id);
+            if (toDelete != null)
+            {
+                _context.PublishRequestDetails.Remove(toDelete);
+                _context.SaveChanges();
+
+                return new GenericResponse<PublishRequestDetail> { Response = toDelete, Code = ResponseCode.OK };
+
+            }
+
+            var toReturn = new GenericResponse<PublishRequestDetail> { Response = toDelete, Code = ResponseCode.BadRequest };
             return toReturn;
         }
